Add a sanity check in database.php

+++ samples/chat/database.php
@@ -1,5 +1,10 @@
 <?php
 
+if ( ! isset($CFG) ) {
+    die("This file is not supposed to be accessed directly.  It is activated using the
+        'Admin' feature from the main page of the application.");
+}
+
 $DATABASE_UNINSTALL = "drop table if exists {$CFG->dbprefix}sample_chat";
 
 $DATABASE_INSTALL = array(
